core code updates
git-svn-id: http://svn.code.sf.net/p/hsqldb/svn/base@2977 7c7dc5f5-a22d-0410-a3af-b41755a11667

+++ trunk/src/org/hsqldb/types/IntervalType.java
@@ -1374,41 +1374,42 @@ public final class IntervalType extends DTIType {
         return limit;
     }
 
-    public long getPart(Session session, Object interval, int part) {
+    public int getPart(Session session, Object interval, int part) {
 
         long units;
 
         switch (part) {
 
             case Types.SQL_INTERVAL_YEAR :
-                return ((IntervalMonthData) interval).units / 12;
+                return (int) (((IntervalMonthData) interval).units / 12);
 
             case Types.SQL_INTERVAL_MONTH :
                 units = ((IntervalMonthData) interval).units;
 
-                return part == startIntervalType ? units
-                                                 : units % 12;
+                return part == startIntervalType ? (int) units
+                                                 : (int) (units % 12);
 
             case Types.SQL_INTERVAL_DAY :
-                return ((IntervalSecondData) interval).units / (24 * 60 * 60);
+                return (int) (((IntervalSecondData) interval).units
+                              / (24 * 60 * 60));
 
             case Types.SQL_INTERVAL_HOUR : {
                 units = ((IntervalSecondData) interval).units / (60 * 60);
 
-                return part == startIntervalType ? units
-                                                 : units % 24;
+                return part == startIntervalType ? (int) units
+                                                 : (int) (units % 24);
             }
             case Types.SQL_INTERVAL_MINUTE : {
                 units = ((IntervalSecondData) interval).units / 60;
 
-                return part == startIntervalType ? units
-                                                 : units % 60;
+                return part == startIntervalType ? (int) units
+                                                 : (int) (units % 60);
             }
             case Types.SQL_INTERVAL_SECOND : {
                 units = ((IntervalSecondData) interval).units;
 
-                return part == startIntervalType ? units
-                                                 : units % 60;
+                return part == startIntervalType ? (int) units
+                                                 : (int) (units % 60);
             }
             default :
                 throw Error.runtimeError(ErrorCode.U_S0500, "IntervalType");
